# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Royal Helmet Quảng Trị storefront and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  productCount: number;
  orderCount: number;
  newOrderCount: number;
  revenue: number;
}