# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Royal Helmet Quảng Trị storefront and admin API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  address: string;
  note?: string;
  total: number;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  items: OrderItem[];
  createdAt: Date;
}