# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Royal Helmet Quảng Trị storefront and admin API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';
import type { OrderLineInput } from './orderLineInput';

export interface OrderInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 8 */
  phone: string;
  /** @minLength 5 */
  address: string;
  note?: string;
  paymentMethod?: OrderInputPaymentMethod;
  /** @minItems 1 */
  items: OrderLineInput[];
}