# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Royal Helmet Quảng Trị storefront and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  /** @nullable */
  size: string | null;
  /** @nullable */
  color: string | null;
}